Adds login_page.dart and corrects variable name.
